requirements.txt
conda create -n py36torch1 python=3.6
pip install absl-py==0.12.0
pip install cachetools==4.2.1
pip install certifi==2020.12.5
pip install chardet==4.0.0
pip install cycler==0.10.0
pip install future==0.18.2
pip install google-auth==1.28.0
pip install google-auth-oauthlib==0.4.4
pip install grpcio==1.36.1
pip install idna==2.10
pip install importlib-metadata==3.10.0
pip install kiwisolver==1.3.1
pip install lmdb==1.1.1
pip install Markdown==3.3.4
pip install matplotlib==3.3.4
pip install numpy==1.19.5
pip install oauthlib==3.1.0
pip install opencv-python==4.5.1.48
pip install pandas==1.1.5
pip install Pillow==8.1.2
pip install pip==21.0.1
pip install protobuf==3.15.6
pip install pyasn1==0.4.8
pip install pyasn1-modules==0.2.8
pip install pyparsing==2.4.7
pip install python-dateutil==2.8.1
pip install pytz==2021.1
pip install PyYAML==5.4.1
pip install requests==2.25.1
pip install requests-oauthlib==1.3.0
pip install rsa==4.7.2
pip install scipy==1.5.4
pip install setuptools==52.0.0.post20210125
pip install six==1.15.0
pip install tensorboard==2.4.1
pip install tensorboard-plugin-wit==1.8.0
pip install tensorboardX==2.1
pip install torch==1.6.0
pip install torchvision==0.7.0
pip install typing-extensions==3.7.4.3
pip install urllib3==1.26.4
pip install Werkzeug==1.0.1
pip install wheel==0.36.2
pip install zipp==3.4.1


Package                Version
---------------------- -------------------
absl-py                0.12.0
cachetools             4.2.1
certifi                2020.12.5
chardet                4.0.0
cycler                 0.10.0
future                 0.18.2
google-auth            1.28.0
google-auth-oauthlib   0.4.4
grpcio                 1.36.1
idna                   2.10
importlib-metadata     3.10.0
kiwisolver             1.3.1
lmdb                   1.1.1
Markdown               3.3.4
matplotlib             3.3.4
numpy                  1.19.5
oauthlib               3.1.0
opencv-python          4.5.1.48
pandas                 1.1.5
Pillow                 8.1.2
pip                    21.0.1
protobuf               3.15.6
pyasn1                 0.4.8
pyasn1-modules         0.2.8
pyparsing              2.4.7
python-dateutil        2.8.1
pytz                   2021.1
PyYAML                 5.4.1
requests               2.25.1
requests-oauthlib      1.3.0
rsa                    4.7.2
scipy                  1.5.4
setuptools             52.0.0.post20210125
six                    1.15.0
tensorboard            2.4.1
tensorboard-plugin-wit 1.8.0
tensorboardX           2.1
torch                  1.6.0
torchvision            0.7.0
typing-extensions      3.7.4.3
urllib3                1.26.4
Werkzeug               1.0.1
wheel                  0.36.2
zipp                   3.4.1
